Unify ACT4 test functions
