--- crates/execution/ab-riscv-act4-runner/src/abundance_rv32i_max/interpreter.rs
@@ -1,164 +1,0 @@
-use crate::abundance_rv32i_max::instruction::AbundanceRv32IMaxInstruction;
-use crate::instruction::{MHPMEVENT3_CSR_INDEX, MHPMEVENT31_CSR_INDEX};
-use ab_riscv_interpreter::prelude::*;
-use ab_riscv_primitives::prelude::*;
-use chacha20::ChaCha8Rng;
-use chacha20::rand_core::{Rng, SeedableRng};
-use std::collections::BTreeMap;
-
-pub(crate) struct AbundanceRv32IMaxExtState {
-    csrs: BTreeMap<u16, u32>,
-    vregs: VectorRegisterFile<const { Self::VLEN }>,
-    reservation: Option<u32>,
-    entropy_source: ChaCha8Rng,
-}
-
-impl AbundanceRv32IMaxExtState {
-    pub(crate) fn new() -> Self {
-        let mut s = Self {
-            csrs: BTreeMap::new(),
-            vregs: VectorRegisterFile::default(),
-            reservation: None,
-            // Good enough for testing purposes
-            entropy_source: ChaCha8Rng::from_seed([0; _]),
-        };
-        // Vector CSRs
-        s.init_csr(VectorCsr::Vstart.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vxsat.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vxrm.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vcsr.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vl.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vtype.to_csr_index(), 1u32 << (u32::BITS - 1));
-        s.init_csr(VectorCsr::Vlenb.to_csr_index(), Self::VLEN.bytes());
-        // Machine trap CSRs - zero-initialized, mtvec must be written by test
-        // boot code before any trap can be taken.
-        s.init_csr(MCsr::Mstatus as u16, 0);
-        // MISA_CSR_IMPLEMENTED is false for this core, so misa is hardwired to 0 - correctly
-        // signaling "no extensions" including no H (bit 7), which the ACT4 framework's own trap
-        // handler (under STANDARD_SM_SUPPORTED) reads unconditionally on every M-mode trap to
-        // decide whether to also save mtval2/mtinst.
-        s.init_csr(MCsr::Misa as u16, 0);
-        s.init_csr(MCsr::Mie as u16, 0);
-        s.init_csr(MCsr::Mtvec as u16, 0);
-        s.init_csr(MCsr::Mscratch as u16, 0);
-        s.init_csr(MCsr::Mepc as u16, 0);
-        s.init_csr(MCsr::Mcause as u16, 0);
-        s.init_csr(MCsr::Mtval as u16, 0);
-        s.init_csr(MCsr::Mip as u16, 0);
-        // mstatush only exists on RV32; the ACT4 framework's default M-mode boot
-        // (RVTEST_BOOT_TO_MMODE, under STANDARD_SM_SUPPORTED) unconditionally clears it
-        // unless SM1P11P0_SUPPORTED is defined, which this core does not define.
-        s.init_csr(MCsr::Mstatush as u16, 0);
-        // mcountinhibit and mhpmevent3..31 are unconditionally zeroed by the same boot
-        // sequence ("They must be implemented" per its own comment) regardless of whether
-        // this core actually implements extra hardware performance counters.
-        s.init_csr(MCsr::Mcountinhibit as u16, 0);
-        for csr_index in MHPMEVENT3_CSR_INDEX..=MHPMEVENT31_CSR_INDEX {
-            s.init_csr(csr_index, 0);
-        }
-        s.init_csr(SEED_CSR_INDEX, 0);
-        s.initialize_vector_state();
-        s
-    }
-
-    fn init_csr(&mut self, idx: u16, val: u32) {
-        self.csrs.insert(idx, val);
-    }
-
-    /// Dispatch a synchronous trap, returning the new PC (mtvec target).
-    ///
-    /// Writes `mepc`, `mcause`, `mtval`, then returns `mtvec & !0b11` (direct mode only -
-    /// MTVEC_MODES: `[0]`). If mtvec is zero, the test never set it up, which means the test
-    /// doesn't expect traps; return None so the caller can treat it as a hard error.
-    pub(crate) fn take_trap<Cause>(&mut self, cause: Cause, epc: u32, tval: u32) -> Option<u32>
-    where
-        MCause: From<Cause>,
-    {
-        *self.csrs.get_mut(&(MCsr::Mepc as u16)).unwrap() = epc;
-        *self.csrs.get_mut(&(MCsr::Mcause as u16)).unwrap() =
-            MCause::from(cause).to_raw::<Reg<u32>>();
-        *self.csrs.get_mut(&(MCsr::Mtval as u16)).unwrap() = tval;
-
-        let mtvec = *self.csrs.get(&(MCsr::Mtvec as u16)).unwrap();
-        if mtvec == 0 {
-            return None;
-        }
-
-        Some(mtvec & !0b11)
-    }
-}
-
-impl Csrs<<AbundanceRv32IMaxInstruction as Instruction>::Reg> for AbundanceRv32IMaxExtState {
-    fn privilege_level(&self) -> PrivilegeLevel {
-        PrivilegeLevel::Machine
-    }
-
-    fn read_csr(&self, csr_index: u16) -> Result<u32, CsrError> {
-        self.csrs
-            .get(&csr_index)
-            .copied()
-            .ok_or(CsrError::IllegalRead { csr_index })
-    }
-
-    fn write_csr(&mut self, csr_index: u16, value: u32) -> Result<(), CsrError> {
-        let slot = self
-            .csrs
-            .get_mut(&csr_index)
-            .ok_or(CsrError::IllegalWrite { csr_index })?;
-        *slot = value;
-        Ok(())
-    }
-}
-
-impl VectorRegistersBase for AbundanceRv32IMaxExtState {
-    const ELEN: Elen = Elen::L64;
-    const VLEN: Vlen = Vlen::L1024;
-}
-
-impl VectorRegisters for AbundanceRv32IMaxExtState
-where
-    Self: Csrs<<AbundanceRv32IMaxInstruction as Instruction>::Reg>,
-{
-    fn read_vregs(&self) -> &VectorRegisterFile<{ Self::VLEN }> {
-        &self.vregs
-    }
-
-    fn write_vregs(&mut self) -> &mut VectorRegisterFile<{ Self::VLEN }> {
-        &mut self.vregs
-    }
-
-    fn vector_instructions_allowed(&self) -> bool {
-        true
-    }
-
-    fn mark_vs_dirty(&mut self) {}
-}
-
-impl VectorRegistersExt<<AbundanceRv32IMaxInstruction as Instruction>::Reg>
-    for AbundanceRv32IMaxExtState
-{
-}
-
-impl ReservationSet<<AbundanceRv32IMaxInstruction as Instruction>::Reg>
-    for AbundanceRv32IMaxExtState
-{
-    fn reservation(&self) -> Option<u32> {
-        self.reservation
-    }
-
-    fn set_reservation(&mut self, address: u32) {
-        self.reservation = Some(address);
-    }
-
-    fn clear_reservation(&mut self) {
-        self.reservation = None;
-    }
-}
-
-impl ZkrSeedSource for AbundanceRv32IMaxExtState {
-    fn poll_seed(&mut self) -> ZkrSeedPoll {
-        let mut randomness = [0; _];
-        self.entropy_source.fill_bytes(&mut randomness);
-        ZkrSeedPoll::Es16(u16::from_le_bytes(randomness))
-    }
-}
