+++ crates/execution/ab-riscv-act4-runner/src/abundance_rv64i_max.rs
@@ -1,2 +1,99 @@
-pub(crate) mod instruction;
-pub(crate) mod interpreter;
+use crate::instruction::MachineModePlaceholder;
+use ab_riscv_interpreter::prelude::*;
+use ab_riscv_macros::{instruction, instruction_execution};
+use ab_riscv_primitives::prelude::*;
+use core::fmt;
+use core::ops::ControlFlow;
+
+/// All instructions supported by the interpreter for RV64I base ISA
+pub(crate) type AbundanceRv64IMaxInstruction = AbundanceRv64IMaxInstructionPrototype<Reg<u64>>;
+
+/// All instructions supported by the interpreter for RV64I base ISA
+#[instruction(
+    inherit = [
+        Rv64Instruction,
+        Rv64AInstruction,
+        Rv64BInstruction,
+        Rv64MInstruction,
+        Rv64ZabhaInstruction,
+        Rv64ZacasInstruction,
+        Rv64ZalasrInstruction,
+        Rv64ZbcInstruction,
+        Rv64ZcaInstruction,
+        Rv64ZcbInstruction,
+        Rv64ZcmpInstruction,
+        Rv64ZknInstruction,
+        ZawrsInstruction,
+        ZicondInstruction,
+        ZicsrInstruction,
+        ZkrInstruction,
+        ZvbbInstruction,
+        ZvbcInstruction,
+        ZveXxInstruction,
+        MachineModePlaceholder,
+    ],
+)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum AbundanceRv64IMaxInstructionPrototype<Reg> {}
+
+#[instruction]
+const impl<Reg> Instruction for AbundanceRv64IMaxInstructionPrototype<Reg> {
+    type Reg = Reg;
+
+    #[inline(always)]
+    fn try_decode(instruction: u32) -> Option<Self> {
+        None
+    }
+
+    #[inline(always)]
+    fn alignment() -> u8 {
+        align_of::<u32>() as u8
+    }
+
+    #[inline(always)]
+    fn size(&self) -> u8 {
+        size_of::<u32>() as u8
+    }
+}
+
+#[instruction]
+impl<Reg> fmt::Display for AbundanceRv64IMaxInstructionPrototype<Reg>
+where
+    Reg: Register,
+{
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {}
+    }
+}
+
+#[instruction_execution]
+impl<Reg> ExecutableInstructionOperands for AbundanceRv64IMaxInstructionPrototype<Reg> {}
+
+#[instruction_execution]
+impl<Reg, ExtState, CustomError> ExecutableInstructionCsr<ExtState, CustomError>
+    for AbundanceRv64IMaxInstructionPrototype<Reg>
+{
+}
+
+#[instruction_execution]
+impl<Reg, Regs, ExtState, Memory, PC, InstructionHandler, CustomError>
+    ExecutableInstruction<Regs, ExtState, Memory, PC, InstructionHandler, CustomError>
+    for AbundanceRv64IMaxInstructionPrototype<Reg>
+where
+    Reg: Register,
+{
+    fn execute(
+        self,
+        Rs1Rs2OperandValues {
+            rs1_value,
+            rs2_value,
+        }: Rs1Rs2OperandValues<<Self::Reg as Register>::Type>,
+        regs: &mut Regs,
+        ext_state: &mut ExtState,
+        memory: &mut Memory,
+        program_counter: &mut PC,
+        system_instruction_handler: &mut InstructionHandler,
+    ) -> ExecutionResult<Self::Reg, CustomError> {
+        ExecutionResult::ContinueNoWrite
+    }
+}
