--- crates/execution/ab-riscv-act4-runner/src/abundance_rv64i_max/interpreter.rs
@@ -1,166 +1,0 @@
-use crate::abundance_rv64i_max::instruction::AbundanceRv64IMaxInstruction;
-use crate::instruction::{MHPMEVENT3_CSR_INDEX, MHPMEVENT31_CSR_INDEX};
-use ab_riscv_interpreter::prelude::*;
-use ab_riscv_primitives::prelude::*;
-use chacha20::ChaCha8Rng;
-use chacha20::rand_core::{Rng, SeedableRng};
-use std::collections::BTreeMap;
-
-pub(crate) struct AbundanceRv64IMaxExtState {
-    csrs: BTreeMap<u16, u64>,
-    vregs: VectorRegisterFile<const { Self::VLEN }>,
-    reservation: Option<u64>,
-    entropy_source: ChaCha8Rng,
-}
-
-impl AbundanceRv64IMaxExtState {
-    pub(crate) fn new() -> Self {
-        let mut s = Self {
-            csrs: BTreeMap::new(),
-            vregs: VectorRegisterFile::default(),
-            reservation: None,
-            // Good enough for testing purposes
-            entropy_source: ChaCha8Rng::from_seed([0; _]),
-        };
-        // Vector CSRs
-        s.init_csr(VectorCsr::Vstart.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vxsat.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vxrm.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vcsr.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vl.to_csr_index(), 0);
-        s.init_csr(VectorCsr::Vtype.to_csr_index(), 1u64 << (u64::BITS - 1));
-        s.init_csr(
-            VectorCsr::Vlenb.to_csr_index(),
-            u64::from(Self::VLEN.bytes()),
-        );
-        // Machine trap CSRs - zero-initialized, mtvec must be written by test
-        // boot code before any trap can be taken.
-        s.init_csr(MCsr::Mstatus as u16, 0);
-        // MISA_CSR_IMPLEMENTED is false for this core, so misa is hardwired to 0 - correctly
-        // signaling "no extensions" including no H (bit 7), which the ACT4 framework's own trap
-        // handler (under STANDARD_SM_SUPPORTED) reads unconditionally on every M-mode trap to
-        // decide whether to also save mtval2/mtinst.
-        s.init_csr(MCsr::Misa as u16, 0);
-        s.init_csr(MCsr::Mie as u16, 0);
-        s.init_csr(MCsr::Mtvec as u16, 0);
-        s.init_csr(MCsr::Mscratch as u16, 0);
-        s.init_csr(MCsr::Mepc as u16, 0);
-        s.init_csr(MCsr::Mcause as u16, 0);
-        s.init_csr(MCsr::Mtval as u16, 0);
-        s.init_csr(MCsr::Mip as u16, 0);
-        // mcountinhibit and mhpmevent3..31 are unconditionally zeroed by the ACT4 framework's
-        // default M-mode boot (RVTEST_BOOT_TO_MMODE, under STANDARD_SM_SUPPORTED) - "They must
-        // be implemented" per its own comment - regardless of whether this core actually
-        // implements extra hardware performance counters. mstatush does not exist on RV64 (its
-        // fields live in mstatus directly), so unlike the RV32 variant there is nothing to add
-        // for it here.
-        s.init_csr(MCsr::Mcountinhibit as u16, 0);
-        for csr_index in MHPMEVENT3_CSR_INDEX..=MHPMEVENT31_CSR_INDEX {
-            s.init_csr(csr_index, 0);
-        }
-        s.init_csr(SEED_CSR_INDEX, 0);
-        s.initialize_vector_state();
-        s
-    }
-
-    fn init_csr(&mut self, idx: u16, val: u64) {
-        self.csrs.insert(idx, val);
-    }
-
-    /// Dispatch a synchronous trap, returning the new PC (mtvec target).
-    ///
-    /// Writes `mepc`, `mcause`, `mtval`, then returns `mtvec & !0b11` (direct mode only -
-    /// MTVEC_MODES: `[0]`). If mtvec is zero, the test never set it up, which means the test
-    /// doesn't expect traps; return None so the caller can treat it as a hard error.
-    pub(crate) fn take_trap<Cause>(&mut self, cause: Cause, epc: u64, tval: u64) -> Option<u64>
-    where
-        MCause: From<Cause>,
-    {
-        *self.csrs.get_mut(&(MCsr::Mepc as u16)).unwrap() = epc;
-        *self.csrs.get_mut(&(MCsr::Mcause as u16)).unwrap() =
-            MCause::from(cause).to_raw::<Reg<u64>>();
-        *self.csrs.get_mut(&(MCsr::Mtval as u16)).unwrap() = tval;
-
-        let mtvec = *self.csrs.get(&(MCsr::Mtvec as u16)).unwrap();
-        if mtvec == 0 {
-            return None;
-        }
-
-        Some(mtvec & !0b11)
-    }
-}
-
-impl Csrs<<AbundanceRv64IMaxInstruction as Instruction>::Reg> for AbundanceRv64IMaxExtState {
-    fn privilege_level(&self) -> PrivilegeLevel {
-        PrivilegeLevel::Machine
-    }
-
-    fn read_csr(&self, csr_index: u16) -> Result<u64, CsrError> {
-        self.csrs
-            .get(&csr_index)
-            .copied()
-            .ok_or(CsrError::IllegalRead { csr_index })
-    }
-
-    fn write_csr(&mut self, csr_index: u16, value: u64) -> Result<(), CsrError> {
-        let slot = self
-            .csrs
-            .get_mut(&csr_index)
-            .ok_or(CsrError::IllegalWrite { csr_index })?;
-        *slot = value;
-        Ok(())
-    }
-}
-
-impl VectorRegistersBase for AbundanceRv64IMaxExtState {
-    const ELEN: Elen = Elen::L64;
-    const VLEN: Vlen = Vlen::L1024;
-}
-
-impl VectorRegisters for AbundanceRv64IMaxExtState
-where
-    Self: Csrs<<AbundanceRv64IMaxInstruction as Instruction>::Reg>,
-{
-    fn read_vregs(&self) -> &VectorRegisterFile<{ Self::VLEN }> {
-        &self.vregs
-    }
-
-    fn write_vregs(&mut self) -> &mut VectorRegisterFile<{ Self::VLEN }> {
-        &mut self.vregs
-    }
-
-    fn vector_instructions_allowed(&self) -> bool {
-        true
-    }
-
-    fn mark_vs_dirty(&mut self) {}
-}
-
-impl VectorRegistersExt<<AbundanceRv64IMaxInstruction as Instruction>::Reg>
-    for AbundanceRv64IMaxExtState
-{
-}
-
-impl ReservationSet<<AbundanceRv64IMaxInstruction as Instruction>::Reg>
-    for AbundanceRv64IMaxExtState
-{
-    fn reservation(&self) -> Option<u64> {
-        self.reservation
-    }
-
-    fn set_reservation(&mut self, address: u64) {
-        self.reservation = Some(address);
-    }
-
-    fn clear_reservation(&mut self) {
-        self.reservation = None;
-    }
-}
-
-impl ZkrSeedSource for AbundanceRv64IMaxExtState {
-    fn poll_seed(&mut self) -> ZkrSeedPoll {
-        let mut randomness = [0; _];
-        self.entropy_source.fill_bytes(&mut randomness);
-        ZkrSeedPoll::Es16(u16::from_le_bytes(randomness))
-    }
-}
